requirements.txt
Flask==0.10.1

Flask-Login==0.2.11
Flask-SQLAlchemy==1.0
Flask-WTF==0.10.0

gevent-socketio==0.3.5-rc2